refactor: native agent construction
