+++ src/agents/native_agents/rec_agent/rec_agent.py
@@ -17,9 +17,10 @@ class RecAgent(BaseAgent):
                  task_input,
                  llm,
                  agent_process_queue,
+                 agent_process_factory,
                  log_mode
         ):
-        BaseAgent.__init__(self, agent_name, task_input, llm, agent_process_queue, log_mode)
+        BaseAgent.__init__(self, agent_name, task_input, llm, agent_process_queue, agent_process_factory, log_mode)
 
     def run(self):
         prompt = ""
