+++ src/utils/utils.py
@@ -10,8 +10,8 @@ import re
 
 import logging
 
-logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
-logger = logging.getLogger(__name__)
+# logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
+# logger = logging.getLogger(__name__)
 
 def parse_global_args():
     parser = argparse.ArgumentParser(description="Parse global parameters")
